requirements.txt
numpy
pymupdf
Pillow
moviepy
sentence-transformers
transformers
torch
python-docx